sys/thread.h: Use native "int" for total_pending.
Saves about 30 Bytes flash on Cortex-M3.

git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6648 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/include/sys/thread.h
@@ -245,7 +245,7 @@ extern NUTTHREADINFO *nutThreadList;
  */
 extern NUTTHREADINFO *runQueue;
 
-extern volatile int16_t total_pending;
+extern volatile int total_pending;
 
 #ifndef __NUT_EMULATION__
 /*!
